requirements.txt
streamlit
pandas
psutil
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
ffmpeg
opencv-python-headless
pillow
numpy
ffmpeg-python
pytube
pytz
google-auth
google-auth-oauthlib
google-api-python-client
requests
google-generativeai
langdetect
langcodes
beautifulsoup4
markdown
